fix(valgrind): detect cached libc debug symbols by file, not dpkg state
The setup cache restores package files onto the root filesystem without
touching dpkg's database, so `dpkg -s libc6-dbg` reported the package as
missing after a cache restore and valgrind's dependencies were reinstalled
on every run.

Probe the system libc for a resolvable separate debug file instead, using
the same build-id and `.gnu_debuglink` lookup GDB and valgrind perform.

The lookup moves from `wall_time::profiler::perf::elf_helper` to a shared
`executor::helpers::debug_file` module; perf's debug-info and module-symbol
resolution now go through it, unchanged. `apt::is_package_installed` had no
other caller and is removed.

Closes COD-3306
